requirements.txt
asgiref==3.7.2
black==23.7.0
boto3==1.28.39
botocore==1.31.39
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
dj-database-url==2.1.0
Django==4.2.4
django-debug-toolbar==4.2.0
django-on-heroku==1.1.2
django-storages==1.13.2
djlint==1.32.1
EditorConfig==0.12.3
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
stripe==7.5.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
django-debug-toolbar